[TASK] Enable some more warnings in the functional tests
Also drop configuration values that are the same as their
corresponding default values.

Do not enable warnings for deprecations (and their output)
yet, though, as we cannot avoid some warnings about usage
of the deprecated `ObjectManager` in TYPO3 11LTS as the
warning is created in Core code. So that part needs to
wait until we have dropped support for TYPO3 11LTS.